+++ src/lib_dal_node_services/types.mli
@@ -255,7 +255,7 @@ val wait_query : < wait : bool > Resto.Query.t
 
 val connected_query : < connected : bool > Resto.Query.t
 
-val subscribed_query : < subscribed : bool > Resto.Query.t
+val all_query : < all : bool > Resto.Query.t
 
 val opt_header_status_query : header_status option Resto.Query.t