+++ tezt/lib_tezos/dal_common.ml
@@ -420,9 +420,9 @@ module Dal_RPC = struct
     make ~query_string:[] GET ["p2p"; "gossipsub"; "topics"] (fun json ->
         JSON.(json |> as_list |> List.map as_topic))
 
-  let get_topics_peers ~subscribed =
+  let get_topics_peers ?(all = false) () =
     let open JSON in
-    let query_string = if subscribed then [("subscribed", "true")] else [] in
+    let query_string = if all then [("all", "true")] else [] in
     let as_topic json =
       let topic_slot_index = get "slot_index" json |> as_int in
       let topic_pkh = get "pkh" json |> as_string in
@@ -436,9 +436,9 @@ module Dal_RPC = struct
     make ~query_string GET ["p2p"; "gossipsub"; "topics"; "peers"] (fun json ->
         JSON.(json |> as_list |> List.map as_topic_and_peers))
 
-  let get_slot_indexes_peers ~subscribed =
+  let get_slot_indexes_peers ?(all = false) () =
     let open JSON in
-    let query_string = if subscribed then [("subscribed", "true")] else [] in
+    let query_string = if all then [("all", "true")] else [] in
     let as_slot_indexes_and_peers json =
       let topic = get "slot_index" json |> as_int in
       let peers = get "peers" json |> as_list |> List.map as_string in
@@ -450,9 +450,9 @@ module Dal_RPC = struct
       ["p2p"; "gossipsub"; "slot_indexes"; "peers"]
       (fun json -> JSON.(json |> as_list |> List.map as_slot_indexes_and_peers))
 
-  let get_pkhs_peers ~subscribed =
+  let get_pkhs_peers ?(all = false) () =
     let open JSON in
-    let query_string = if subscribed then [("subscribed", "true")] else [] in
+    let query_string = if all then [("all", "true")] else [] in
     let as_pkhs_and_peers json =
       let topic = get "pkh" json |> as_string in
       let peers = get "peers" json |> as_list |> List.map as_string in